refactor: 단건, 다건 분류를 위해 패키지 위치 이동
