+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/notion/infrastructure/single/SyncedPagesDataRepository.java
@@ -1,4 +1,4 @@
-package io.f12.notionlinkedblog.notion.infrastructure;
+package io.f12.notionlinkedblog.notion.infrastructure.single;
 
 import java.util.List;
 import java.util.Optional;
