+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/notion/service/port/SyncedPagesRepository.java
@@ -3,7 +3,7 @@ package io.f12.notionlinkedblog.notion.service.port;
 import java.util.List;
 import java.util.Optional;
 
-import io.f12.notionlinkedblog.notion.infrastructure.SyncedPagesEntity;
+import io.f12.notionlinkedblog.notion.infrastructure.single.SyncedPagesEntity;
 
 public interface SyncedPagesRepository {
 	List<SyncedPagesEntity> findAll();
